As we will be moving the playlist functionality to the status page, give it a real server.
