+++ status/index.js
@@ -1,138 +1,51 @@
-/* Copyright 2019 Google Inc. All Rights Reserved.
+import commandLineArgs from 'command-line-args';
+import commandLineUsage from 'command-line-usage';
+import express from 'express';
+import fs from 'fs';
+import path from 'path';
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-==============================================================================*/
-
-import {ClientController} from './client_controller.js';
-import {PlaylistController} from './playlist_controller.js';
-import {ErrorController} from './error_controller.js';
-import {PlaylistCreator} from './playlist_creator.js';
-import io from './socket.io-client.js';
-
-let lastUpdateFromServer = 0;
-let timeOfLastUpdateFromServer = window.performance.now();
-let connected = false;
-function getTime() {
-  if (!connected) {
-    return lastUpdateFromServer;
-  }
-  return lastUpdateFromServer + window.performance.now() - timeOfLastUpdateFromServer;
+const FLAG_DEFS = [
+  {name: 'port', type: Number, defaultValue: 3000}
+];
+const flags = commandLineArgs(FLAG_DEFS);
+if (flags.help) {
+  console.log('Available flags: ' + commandLineUsage({optionList: FLAG_DEFS}));
+  process.exit();
 }
 
-const control = io('http://localhost:3000/control');
-const creatorEl = document.querySelector('#playlist-creator');
+const fsp = fs.promises;
 
-function applyNewPlaylist(playlist, moduleConfig) {
-  // TODO(applmak): Passing a string here is a bit hacky.
-  if (playlist == 'reset') {
-    control.emit('resetPlaylist');
-  } else {
-    control.emit('newPlaylist', {playlist, moduleConfig});
-  }
+function serveFile(path) {
+  return async (req, res) => {
+    try {
+      const contents = await fsp.readFile(path, {encoding: 'utf-8'});
+      res.statusCode = 200;
+      res.end(contents, 'utf-8');
+    } catch (e) {
+      res.statusCode = 404;
+      res.end('Not Found', 'utf-8');
+    }
+  };
 }
 
-const playlistCreator = new PlaylistCreator(creatorEl, applyNewPlaylist);
-const playlistController = new PlaylistController(document.querySelector('.playlist-container'), getTime);
-const errorController = new ErrorController(document.querySelector('footer'));
-const clientController = new ClientController(
-  document.querySelector('.diagram'),
-  req => control.emit('takeSnapshot', req),
-  errorController,
-  getTime,
-);
+// The location we are running from.
+const cwd = path.join(process.cwd(), 'status');
 
-function convertMsDurationToText(ms) {
-  const secs = Math.floor(ms / 1000);
-  const mins = Math.floor(secs / 60);
+console.log(`CWD: ${cwd}`);
 
-  return mins ? `${mins} minutes` : secs ? `${secs} seconds` : String(ms);
-}
+// Create a router just for the brick files that could be served to the client.
+// These are:
+//   /client => node_modules/brick/client
+//   /lib => node_modules/brick/lib
+//   /node_modules => node_modules
+const app = express();
+app.use('/node_modules', express.static(path.join(cwd, 'node_modules')));
+app.use('/js', express.static(path.join(cwd, 'static')));
+app.get('/', serveFile(path.join(cwd, 'static/index.html')));
 
-let transitionData = {};
-control.on('transition', data => {
-  transitionData = data;
+const server = app.listen(flags.port, () => {
+  const host = server.address().address;
+  const port = server.address().port;
 
-  const duration = data.nextDeadline - data.deadline;
-  const moduleNameEl = document.querySelector('#module');
-  moduleNameEl.textContent = data.module;
-  const durationEl = document.querySelector('#duration');
-  durationEl.textContent = convertMsDurationToText(duration);
-
-  playlistController.updateTransitionData(data);
-  playlistCreator.setLivePlaylist(data.layouts);
-  playlistCreator.setModuleConfig(data.configMap);
+  console.log(`Server listening at http://${host}:${port}`);
 });
-control.on('clients', data => {
-  clientController.setClients(data);
-})
-control.on('connect', () => {
-  connected = true;
-  document.querySelector('#disconnected-warning').style.visibility = 'hidden';
-});
-control.on('disconnect', () => {
-  connected = false;
-  document.querySelector('#disconnected-warning').style.visibility = 'visible';
-  playlistController.disconnect();
-  errorController.disconnect();
-  clientController.disconnect();
-});
-control.on('time', data => {
-  lastUpdateFromServer = data.time;
-  timeOfLastUpdateFromServer = window.performance.now();
-});
-control.on('error', e => {
-  playlistController.error(e);
-  errorController.error(e);
-});
-control.on('errors', es => {
-  es.forEach(e => {
-    errorController.error(e);
-  });
-});
-control.on('new-client', c => {
-  clientController.newClient(c);
-});
-control.on('lost-client', c => {
-  clientController.lostClient(c);
-});
-control.on('wallGeometry', p => {
-  clientController.setWallGeometry(p);
-});
-control.on('takeSnapshotRes', res => {
-  clientController.takeSnapshotRes(res);
-});
-
-const openCreatorEl = document.querySelector('#open-creator');
-openCreatorEl.addEventListener('click', () => {
-  playlistCreator.open();
-});
-
-const timeEl = document.querySelector('#time');
-const remainingEl = document.querySelector('#remaining');
-function render() {
-  timeEl.textContent = getTime().toFixed(0);
-  const remainingMs = transitionData.nextDeadline - getTime();
-  if (remainingMs < 0) {
-    remainingEl.textContent = `Fading (${-remainingMs})`;
-    remainingEl.classList.add('transitioning');
-  } else {
-    remainingEl.classList.remove('transitioning');
-    remainingEl.textContent = convertMsDurationToText(remainingMs);
-  }
-
-
-  playlistController.render();
-
-  window.requestAnimationFrame(render);
-}
-render();
