+++ status/package.json
@@ -8,5 +8,6 @@
     "http-server": "^0.11.1",
     "socket.io-client": "^2.3.0",
     "sortablejs": "^1.10.0"
-  }
+  },
+  "type": "module"
 }
