+++ status/static/client_controller.js
