+++ status/static/index.html
@@ -290,7 +290,7 @@
       display: none;
     }
     </style>
-    <script src="./index.js" type="module"></script>
+    <script src="/js/index.js" type="module"></script>
   </head>
   <body>
     <header>
